refactor: extract block edit utils
