+++ packages/crepe/src/feature/block-edit/menu/config.ts
@@ -1,5 +1,3 @@
-import type { html } from 'atomico'
-import type { Ctx } from '@milkdown/kit/ctx'
 import { editorViewCtx } from '@milkdown/kit/core'
 import {
   blockquoteSchema,
@@ -11,9 +9,6 @@ import {
   orderedListSchema,
   paragraphSchema,
 } from '@milkdown/kit/preset/commonmark'
-import type { Attrs, NodeType } from '@milkdown/kit/prose/model'
-import { findWrapping } from '@milkdown/kit/prose/transform'
-import type { Command, Transaction } from '@milkdown/kit/prose/state'
 import { NodeSelection } from '@milkdown/kit/prose/state'
 import { imageBlockSchema } from '@milkdown/kit/component/image-block'
 import { createTable } from '@milkdown/kit/preset/gfm'
@@ -35,90 +30,8 @@ import {
   todoListIcon,
 } from '../../../icons'
 import type { BlockEditFeatureConfig } from '../index'
-
-interface MenuItem {
-  index: number
-  key: string
-  label: string
-  icon: ReturnType<typeof html>
-  onRun: (ctx: Ctx) => void
-}
-
-type WithRange<T, HasIndex extends true | false = true> = HasIndex extends true ? T & { range: [start: number, end: number] } : T
-
-type MenuItemGroup<HasIndex extends true | false = true> = WithRange<{
-  key: string
-  label: string
-  items: HasIndex extends true ? MenuItem[] : Omit<MenuItem, 'index'>[]
-}, HasIndex>
-
-function clearRange(tr: Transaction) {
-  const { $from, $to } = tr.selection
-  const { pos: from } = $from
-  const { pos: to } = $to
-  tr = tr.deleteRange(from - $from.node().content.size, to)
-  return tr
-}
-
-function setBlockType(tr: Transaction, nodeType: NodeType, attrs: Attrs | null = null) {
-  const { from, to } = tr.selection
-  return tr.setBlockType(from, to, nodeType, attrs)
-}
-
-function wrapInBlockType(tr: Transaction, nodeType: NodeType, attrs: Attrs | null = null) {
-  const { $from, $to } = tr.selection
-
-  const range = $from.blockRange($to)
-  const wrapping = range && findWrapping(range, nodeType, attrs)
-  if (!wrapping)
-    return null
-
-  return tr.wrap(range, wrapping)
-}
-
-function addBlockType(tr: Transaction, nodeType: NodeType, attrs: Attrs | null = null) {
-  const node = nodeType.createAndFill(attrs)
-  if (!node)
-    return null
-
-  return tr.replaceSelectionWith(node)
-}
-
-function clearContentAndSetBlockType(nodeType: NodeType, attrs: Attrs | null = null): Command {
-  return (state, dispatch) => {
-    if (dispatch) {
-      const tr = setBlockType(clearRange(state.tr), nodeType, attrs)
-      dispatch(tr.scrollIntoView())
-    }
-    return true
-  }
-}
-
-function clearContentAndWrapInBlockType(nodeType: NodeType, attrs: Attrs | null = null): Command {
-  return (state, dispatch) => {
-    const tr = wrapInBlockType(clearRange(state.tr), nodeType, attrs)
-    if (!tr)
-      return false
-
-    if (dispatch)
-      dispatch(tr.scrollIntoView())
-
-    return true
-  }
-}
-
-function clearContentAndAddBlockType(nodeType: NodeType, attrs: Attrs | null = null): Command {
-  return (state, dispatch) => {
-    const tr = addBlockType(clearRange(state.tr), nodeType, attrs)
-    if (!tr)
-      return false
-
-    if (dispatch)
-      dispatch(tr.scrollIntoView())
-
-    return true
-  }
-}
+import type { MenuItemGroup } from './utils'
+import { clearContentAndAddBlockType, clearContentAndSetBlockType, clearContentAndWrapInBlockType, clearRange } from './utils'
 
 export function getGroups(filter?: string, config?: BlockEditFeatureConfig) {
   let groups: MenuItemGroup<false>[] = [
